+++ src/main/java/com/lattice/controller/TodoListController.java
@@ -9,6 +9,7 @@ import com.lattice.service.TodoItemService;
 import com.lattice.service.TodoListService;
 import org.json.simple.JSONObject;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -27,28 +28,42 @@ public class TodoListController {
     TodoListRepository listRepository;
 
     @PostMapping(value = "/list")
-    public ResponseEntity<String > newList(@RequestBody TodoList list) {
+    public ResponseEntity<String> newList(@RequestBody TodoList list) {
         TodoList createdList = listService.createList(list);
         JSONObject response = new JSONObject();
         response.put("message", "list created successfully");
-        response.put("listId" ,createdList.getListId());
+        response.put("listId", createdList.getListId());
         return ResponseEntity.ok(response.toString());
     }
 
     @PutMapping("/list")
     public ResponseEntity<?> editList(@RequestBody TodoList list) {
-        listService.editList(list);
         JSONObject response = new JSONObject();
-        response.put("message", "list edited successfully");
-        return ResponseEntity.ok(response.toString());
+        boolean exists = listRepository.existsTodoListByListId(list.getListId());
+        if (!exists) {
+            response.put("message", "list doesn't exists");
+            return new ResponseEntity(response.toString(), HttpStatus.NOT_IMPLEMENTED);
+        } else {
+            listService.editList(list);
+            response.put("message", "list edited successfully");
+            return ResponseEntity.ok(response.toString());
+
+        }
     }
 
     @DeleteMapping("/list/{id}")
-    public ResponseEntity<String> deleteList(@PathVariable Long id){
-        listService.deleteList(id);
+    public ResponseEntity<String> deleteList(@PathVariable Long id) {
+
         JSONObject response = new JSONObject();
-        response.put("message", "list deleted successfully");
-        return ResponseEntity.ok(response.toString());
+        boolean exists = listRepository.existsTodoListByListId(id);
+        if (!exists) {
+            response.put("message", "list doesn't exists");
+            return new ResponseEntity(response.toString(), HttpStatus.NOT_IMPLEMENTED);
+        }else {
+            listService.deleteList(id);
+            response.put("message", "list deleted successfully");
+            return ResponseEntity.ok(response.toString());
+        }
     }
 
 
@@ -57,6 +72,7 @@ public class TodoListController {
     public List<TodoItem> getItemsInList(@PathVariable long listId) {
         return itemService.getAllTodoItemsForListId(listId);
     }
+
     // Get todo list, based on listId
     @GetMapping("/list")
     public List<TodoList> getAllLists() {
